added features: - added vue-router - created components Todos, Home - refactored App.vue - created router and configured lazy loading - created routes '/', '/todos' and corresponding views - added homepage
