+++ src/main.js
@@ -1,5 +1,6 @@
 import Vue from 'vue'
 import App from './App.vue'
+import router from './router'
 
 Vue.config.productionTip = false
 
@@ -10,6 +11,7 @@ Vue.filter('formatDate', function(value) {
 })
 
 new Vue({
+  router,
   render: h => h(App),
 }).$mount('#app')
