+++ src/views/Home.vue
@@ -1,0 +1,18 @@
+<template>
+  <div>
+    <h2>Home page</h2>
+    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Delectus doloremque dolorum exercitationem nemo nulla
+      officiis placeat quibusdam quos recusandae voluptas.</p>
+    <router-link to="/todos">Todos</router-link>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "Home"
+}
+</script>
+
+<style scoped>
+
+</style>
